Create play.js
I deleted all of the other bot commands such as "pause" "resume" "stop" & "help". I deleted the files because I wanted to disable the need for a YouTube url. I couldn't figure out how to let the bot run on its own without a url link though. This code above shows the failure of my efforts to disable the need for a url. I coded this back in 2018 when I was just starting out. I thought I might share it.